+++ esma.cmake
@@ -33,6 +33,13 @@ option(BUILD_SHARED_LIBS "Build the shared library" OFF)
 set (ECBUILD_2_COMPAT_VALUE OFF)
 include (ecbuild_system NO_POLICY_SCOPE)
 
+### GPU Physics ###
+
+option(GPU_PHYSICS "Build GPU ports of Physics components" OFF)
+if (GPU_PHYSICS)
+  find_package (OpenACC)
+endif ()
+
 ### Compiler Support ###
 
 list (APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/compiler")